Extract upgrade feature card enum
